Add toestr message for category

+++ public/modules/shopAdmin/controllers/category/category.edit.controller.js
@@ -1,7 +1,7 @@
 'use strict';
 
-angular.module('shopAdmin').controller('categoryEditController', ['$scope', '$state', '$stateParams', 'categoryService',
-    function($scope, $state, $stateParams, categoryService) {
+angular.module('shopAdmin').controller('categoryEditController', ['$scope', '$window', '$state', '$stateParams', 'categoryService',
+    function($scope, $window, $state, $stateParams, categoryService) {
         $scope.categoryId = $stateParams.categoryId;
         $scope.category = {};
         $scope.category.meta = {};
@@ -38,15 +38,19 @@ angular.module('shopAdmin').controller('categoryEditController', ['$scope', '$st
             categoryService.updateCategory($scope.category)
                 .$promise
                 .then(function(response) {
-                   $state.go('Category.List');
+                    $window.toastr.success('Updated category');
+                    $state.go('Category.List');
                 });
         };
 
         $scope.delete = function () {
-            categoryService.deleteCategory($scope.category._id)
-                .$promise.then(function(promise) {
+            if(confirm('Are you sure want to delete this category')) {
+                categoryService.deleteCategory($scope.category._id)
+                    .$promise.then(function(promise) {
+                    $window.toastr.success('Deleted category');
                     $state.go('Category.List');
                 });
+            }
         };
     }
 ]);
